+++ CHANGELOG.md
@@ -20,6 +20,8 @@ best friend, lajbel, can put the correct version name here
 
 ### Added
 
+- Added `getSpriteOutline()` that takes a sprite asset and
+  returns a polygon showing the outline.
 - Added parameterized formatting tags like `"[color=red]Red text![/color]"` in
   `CharTransformFunc` for more powerful text formatting options -
   @dragoncoder047